alsa-lib: 1.1.9 -> 1.2.1.2
Changelogs:
https://alsa-project.org/wiki/Changes_v1.1.9_v1.2.1
https://alsa-project.org/wiki/Changes_v1.2.1_v1.2.1.1
https://alsa-project.org/wiki/Changes_v1.2.1.1_v1.2.1.2

This upgrade brings support for the new "UCM2" configuration format.
Another big change is that the UCM and topology configuration files were
moved to separate repositories, so those are removed from alsa-lib and
will need to be packaged separately.

The topology API was moved from libasound to a separate libatopology
library. This seems like a recipe for regressions for applications that
used to use libasound for the topology stuff, but I don't think there
are many applications in existence that use the topology API. I hope the
upstream knows what they are doing.

Patches added:

0001-configure.ac-remove-an-unnecessary-libtool-fix.patch
This fixes a failure during configure.

0001-ucm-Use-strncmp-to-avoid-access-out-of-boundary.patch
0002-ucm-return-always-at-least-NULL-if-no-list-is-availa.patch
Backports. I don't know if these are regression fixes, but since the UCM
code was heavily worked on in this upgrade, it seems likely that these
patches fix regressions.

0003-namehint-correct-the-args-check.patch
0004-namehint-improve-the-previous-patch-check-the-return.patch
Backports. A change in alsa-plugins caused a regression in the pulse
plugin, but that was actually due to a bug in alsa-lib, which is fixed
by these patches.

0005-ucm-Do-not-fail-to-parse-configs-on-cards-with-an-em.patch
Backport. Fixes a regression with some Intel hardware.

0001-Fix-alsa-sound-.h-for-external-programs.patch
0001-uapi-Move-typedefs-from-uapi-to-sound.patch
Backports. These fix build failures in alsa-tools.

I removed alsa-doc from PACKAGES, because no files were assigned to it.

Signed-off-by: Tanu Kaskinen <tanuk@iki.fi>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>